Splitted Toolbar class from StockToolsGui.
